requirements.txt
streamlit
pandas
openpyxl
langchain
plotly
python-dotenv
google-generativeai
langchain-google-genai